Changes to the Viterbi algorithm. Now scaled and seems to be working fine
